requirements.txt
torch
numpy==1.26
gsplat==0.1.11
nerfstudio==1.1.0
scipy
matplotlib
imageio
tensorboard
pyyaml
future
scikit-learn
opencv-python
tqdm
plyfile
ninja
moviepy
psutil
pdbpp
natsort
trimesh
nvtx
pymcubes
dataclass-wizard[yaml]
torch_efficient_distloss
point-cloud-utils
vector-quantize-pytorch
einops == 0.7.0
deepdish @ git+https://github.com/uchicago-cs/deepdish@b4f399a 
e3nn
-e deps/fast-gca
-e deps/polyscope-py
-e .